requirements.txt
django
django-cors-headers
djangorestframework
python-dateutil
scipy